# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Mavericks Inventory API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /stocks:
    get:
      operationId: listStocks
      tags: [stocks]
      summary: List all stock items
      parameters:
        - in: query
          name: category
          schema: { type: string }
        - in: query
          name: location
          schema: { type: string }
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: search
          schema: { type: string }
      responses:
        "200":
          description: List of stock items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StockItem"
    post:
      operationId: createStock
      tags: [stocks]
      summary: Create a single stock item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StockItemInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockItem"
  /stocks/{id}:
    get:
      operationId: getStock
      tags: [stocks]
      summary: Get a stock item by ID
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Stock item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockItem"
        "404":
          description: Not found
    patch:
      operationId: updateStock
      tags: [stocks]
      summary: Update a stock item
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StockItemUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockItem"
    delete:
      operationId: deleteStock
      tags: [stocks]
      summary: Soft delete a stock item
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
  /stocks/{id}/ledger:
    get:
      operationId: getStockLedger
      tags: [stocks]
      summary: Get stock movement ledger for an item
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Stock ledger entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LedgerEntry"

  # ── Distributions ─────────────────────────────────────────────
  /distributions:
    get:
      operationId: listDistributions
      tags: [distributions]
      summary: List distribution entries
      parameters:
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: stockId
          schema: { type: integer }
      responses:
        "200":
          description: Distribution list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Distribution"
    post:
      operationId: createDistribution
      tags: [distributions]
      summary: Create a distribution entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DistributionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Distribution"
  /distributions/{id}:
    get:
      operationId: getDistribution
      tags: [distributions]
      summary: Get a distribution entry
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Distribution entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Distribution"
        "404":
          description: Not found
    patch:
      operationId: updateDistribution
      tags: [distributions]
      summary: Update a distribution entry (Draft or Rejected only)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DistributionUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Distribution"
    delete:
      operationId: deleteDistribution
      tags: [distributions]
      summary: Delete a distribution entry (Draft only)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
  /distributions/{id}/submit:
    post:
      operationId: submitDistribution
      tags: [distributions]
      summary: Submit a distribution for approval
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Distribution"

  # ── Approvals ─────────────────────────────────────────────────
...
components:
  schemas:
...
    StockItem:
      type: object
      required: [id, stockCode, stockName, category, unitOfMeasure, openingQuantity, availableQuantity, location, status, createdAt]
      properties:
        id:              { type: integer }
        stockCode:       { type: string }
        stockName:       { type: string }
        category:        { type: string }
        unitOfMeasure:   { type: string }
        openingQuantity: { type: number }
        availableQuantity: { type: number }
        minStockLevel:   { type: ["number", "null"] }
        location:        { type: ["string", "null"] }
        status:          { type: string }
        healthScore:     { type: ["number", "null"] }
        createdAt:       { type: string }
        updatedAt:       { type: ["string", "null"] }
    StockItemInput:
      type: object
      required: [stockCode, stockName, category, unitOfMeasure, openingQuantity]
      properties:
        stockCode:       { type: string }
        stockName:       { type: string }
        category:        { type: string }
        unitOfMeasure:   { type: string }
        openingQuantity: { type: number }
        minStockLevel:   { type: number }
        location:        { type: string }
        status:          { type: string }
    StockItemUpdate:
      type: object
      properties:
        stockName:       { type: string }
        category:        { type: string }
        unitOfMeasure:   { type: string }
        location:        { type: string }
        minStockLevel:   { type: number }
        status:          { type: string }
    LedgerEntry:
      type: object
      required: [id, stockId, type, quantity, description, createdAt]
      properties:
        id:          { type: integer }
        stockId:     { type: integer }
        type:        { type: string }
        quantity:    { type: number }
        description: { type: string }
        createdAt:   { type: string }
        createdBy:   { type: ["string", "null"] }
    Distribution:
      type: object
      required: [id, stockId, stockName, quantityDistributed, distributionDate, recipient, purpose, status, createdAt]
      properties:
        id:                   { type: integer }
        stockId:              { type: integer }
        stockName:            { type: string }
        stockCode:            { type: ["string", "null"] }
        quantityDistributed:  { type: number }
        distributionDate:     { type: string }
        recipient:            { type: string }
        purpose:              { type: ["string", "null"] }
        status:               { type: string }
        remarks:              { type: ["string", "null"] }
        submittedAt:          { type: ["string", "null"] }
        approvedAt:           { type: ["string", "null"] }
        createdAt:            { type: string }
    DistributionInput:
      type: object
      required: [stockId, quantityDistributed, distributionDate, recipient]
      properties:
        stockId:              { type: integer }
        quantityDistributed:  { type: number }
        distributionDate:     { type: string }
        recipient:            { type: string }
        purpose:              { type: string }
    DistributionUpdate:
      type: object
      properties:
        quantityDistributed:  { type: number }
        distributionDate:     { type: string }
        recipient:            { type: string }
        purpose:              { type: string }
...