# lib/api-zod/src/generated/types/distributionInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Mavericks Inventory API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DistributionInput {
  stockId: number;
  quantityDistributed: number;
  distributionDate: string;
  recipient: string;
  purpose?: string;
}